+++ src/rpc/mod.rs
@@ -176,8 +176,10 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::state::StateVerifierStatus);
         $callback!(crate::rpc::state::StateGetClaim);
         $callback!(crate::rpc::state::StateGetClaims);
+        $callback!(crate::rpc::state::StateGetAllClaims);
         $callback!(crate::rpc::state::StateGetAllocation);
         $callback!(crate::rpc::state::StateGetAllocations);
+        $callback!(crate::rpc::state::StateGetAllAllocations);
         $callback!(crate::rpc::state::StateGetAllocationIdForPendingDeal);
         $callback!(crate::rpc::state::StateGetAllocationForPendingDeal);
         $callback!(crate::rpc::state::StateSectorExpiration);
@@ -324,7 +326,6 @@ const MAX_RESPONSE_BODY_SIZE: u32 = MAX_REQUEST_BODY_SIZE;
 /// data.
 pub struct RPCState<DB> {
     pub keystore: Arc<RwLock<KeyStore>>,
-    pub chain_store: Arc<crate::chain::ChainStore<DB>>,
     pub state_manager: Arc<crate::state_manager::StateManager<DB>>,
     pub mpool: Arc<crate::message_pool::MessagePool<crate::message_pool::MpoolRpcProvider<DB>>>,
     pub bad_blocks: Arc<crate::chain_sync::BadBlockCache>,
@@ -333,13 +334,28 @@ pub struct RPCState<DB> {
     pub network_name: String,
     pub tipset_send: flume::Sender<Arc<Tipset>>,
     pub start_time: chrono::DateTime<chrono::Utc>,
-    pub beacon: Arc<crate::beacon::BeaconSchedule>,
     pub shutdown: mpsc::Sender<()>,
 }
 
 impl<DB: Blockstore> RPCState<DB> {
+    pub fn beacon(&self) -> &Arc<crate::beacon::BeaconSchedule> {
+        self.state_manager.beacon_schedule()
+    }
+
+    pub fn chain_store(&self) -> &Arc<crate::chain::ChainStore<DB>> {
+        self.state_manager.chain_store()
+    }
+
+    pub fn chain_index(&self) -> &Arc<crate::chain::index::ChainIndex<Arc<DB>>> {
+        &self.chain_store().chain_index
+    }
+
+    pub fn chain_config(&self) -> &Arc<crate::networks::ChainConfig> {
+        self.state_manager.chain_config()
+    }
+
     pub fn store(&self) -> &DB {
-        self.chain_store.blockstore()
+        self.chain_store().blockstore()
     }
 
     pub fn store_owned(&self) -> Arc<DB> {